componentized home page
